requirements.txt
botbuilder-integration-aiohttp>=4.14.0
botbuilder-core
asyncio
aiohttp
fastapi
pydantic
uvicorn
openai
llama_index
websocket

# comment out the line below if you do not have a GPU
--index-url https://download.pytorch.org/whl/cu118
torch
torchvision
torchaudio 
